Improve manual setup
